Eager load all things for api/product/:id
HOYABOYA NOTES: Had to unddo the conflict on products_controller and spec to defer to the version we had in 2-1-dev. It was very very different from Spree master for def update.

https://github.com/bonobos/spree/blob/2-1-dev/api/app/controllers/spree/api/products_controller.rb#L40-L48

Need to call `reload` on @product at update action so that it loads new
variants and show them properly on the template

Fixes #4287
Fixes #4285

Conflicts:
	api/app/controllers/spree/api/products_controller.rb
	api/spec/controllers/spree/api/products_controller_spec.rb