Sign in the current API user when authenticating from Token
Currently, when authenticating from token, we have to use
`@current_api_user`. This will allow any code that relies on
`spree_current_user` or `try_spree_current_user` to continue to work
correctly. As we use `store: false`, this will only apply to the current
request and not attempt to store the auth in the session.