Fixes incorrect variant images being displayed for certain variants
Currently when a variant is supplied to one of the generated image helpers and the variant does not have any associated images, the first of the variant’s product’s variant_images is used. The result is that incorrect images are displayed for certain variants. This can be misleading to admins since the image isn’t incorrect, it’s just not present. This change will display a “noimage” image when a variant is supplied to the image helper methods and it has no associated images.